backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
greenlet>=3.0.0
aiosqlite==0.20.0
pydantic>=2.11.7,<3.0.0
pydantic-settings>=2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
reportlab==4.2.5
Pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy>=1.24.0
